renamed server/routes to server/handlers
