+++ server/handlers/add-new-module.js
