+++ server/handlers/join-module.js
