+++ server/handlers/validate-module.js
