+++ server/routes.js
@@ -1,15 +1,16 @@
-var Index = require('./routes/index');
-var GetModuleList = require('./routes/get-module-list');
-var GetModule = require('./routes/get-module');
-var ValidateModule = require('./routes/validate-module');
-var AddNewModule = require('./routes/add-new-module');
-var SaveUser = require('./routes/save-user');
-var ValidateUser = require('./routes/authenticate-user');
-var GetUserDetails = require('./routes/get-user-details');
-var SaveQuiz = require('./routes/save-quiz');
-var GetQuizQuestions = require('./routes/get-quiz-questions');
-var JoinModule = require('./routes/join-module');
-var SaveStudentResponse = require('./routes/save-student-response');
+var Index = require('./handlers/index');
+var GetModuleList = require('./handlers/get-module-list');
+var GetModule = require('./handlers/get-module');
+var ValidateModule = require('./handlers/validate-module');
+var AddNewModule = require('./handlers/add-new-module');
+var SaveUser = require('./handlers/save-user');
+var ValidateUser = require('./handlers/authenticate-user');
+var GetUserDetails = require('./handlers/get-user-details');
+var SaveQuiz = require('./handlers/save-quiz');
+var GetQuizQuestions = require('./handlers/get-quiz-questions');
+var JoinModule = require('./handlers/join-module');
+var SaveStudentResponse = require('./handlers/save-student-response');
+var EndQuiz = require('./handlers/end-quiz');
 
 var routes = [
     Index,
@@ -23,7 +24,8 @@ var routes = [
     SaveQuiz,
     GetQuizQuestions,
     JoinModule,
-    SaveStudentResponse
+    SaveStudentResponse,
+    EndQuiz
 ];
 
 module.exports = routes;
